MIR-1657 make MIRInstitutes the interface and MIRSampleInstitutes the enum
