+++ mir-it/src/test/java/org/mycore/mir/it/model/MIRInstitutes.java
@@ -1,18 +1,6 @@
 package org.mycore.mir.it.model;
 
-public enum MIRInstitutes implements MIRSampleInstitutes {
-    Universität_in_Deutschland("Unis"),
-    Universität_Jena("Unis.Jena"),
-    IBM("Firma.IBM");
+public interface MIRInstitutes {
 
-    private String value;
-
-    MIRInstitutes(String value) {
-        this.value = value;
-    }
-
-    @Override
-    public String getValue() {
-        return value;
-    }
+    String getValue();
 }
