+++ mir-it/src/test/java/org/mycore/mir/it/tests/MIRAuthorEditorITCase.java
@@ -465,8 +465,8 @@ public class MIRAuthorEditorITCase extends MIRITBase {
         return "MODS-Dokument erstellen";
     }
 
-    protected MIRSampleInstitutes institutionTestValue() {
-        return MIRInstitutes.Universität_in_Deutschland;
+    protected MIRInstitutes institutionTestValue() {
+        return MIRSampleInstitutes.Universität_in_Deutschland;
     }
 
     protected String institutionValidationText() {
